Lots of updates

+++ lib/MPU9250/kalman.h
@@ -5,17 +5,18 @@
 #ifndef _Kalman_h
 #define _Kalman_h
 
+
 class Kalman {
   private:
     /* Kalman filter variables */
-    double q; //process noise covariance
-    double r; //measurement noise covariance
-    double x; //value
-    double p; //estimation error covariance
-    double k; //kalman gain
+    float q; //process noise covariance
+    float r; //measurement noise covariance
+    float x; //value
+    float p; //estimation error covariance
+    float k; //kalman gain
     
   public:
-    Kalman(double process_noise, double sensor_noise, double estimated_error, double intial_value) {
+    Kalman(float process_noise, float sensor_noise, float estimated_error, float intial_value) {
       /* The variables are x for the filtered value, q for the process noise, 
          r for the sensor noise, p for the estimated error and k for the Kalman Gain. 
          The state of the filter is defined by the values of these variables.
@@ -40,7 +41,8 @@ class Kalman {
         this->x = intial_value; //x will hold the iterated filtered value
     }
     
-    double getFilteredValue(double measurement) {
+    float getFilteredValue(float measurement) 
+    {
       /* Updates and gets the current measurement value */
       //prediction update
       //omit x = x
@@ -54,26 +56,26 @@ class Kalman {
       return this->x;
     }
     
-    void setParameters(double process_noise, double sensor_noise, double estimated_error) {
+    void setParameters(float process_noise, float sensor_noise, float estimated_error) {
         this->q = process_noise;
         this->r = sensor_noise;
         this->p = estimated_error;
     }
 
-    void setParameters(double process_noise, double sensor_noise) {
+    void setParameters(float process_noise, float sensor_noise) {
         this->q = process_noise;
         this->r = sensor_noise;
     }
     
-    double getProcessNoise() {
+    float getProcessNoise() {
       return this->q;
     }
     
-    double getSensorNoise() {
+    float getSensorNoise() {
       return this->r;
     }
     
-    double getEstimatedError() {
+    float getEstimatedError() {
       return this->p;
     }
 };
